refactor : add register api file, refactor article api file's name - reason: for server-side API, refactored article api file's name to route
