+++ src/components/ArticleContent/index.tsx
@@ -3,7 +3,7 @@
 import ReactQuill from 'react-quill'
 import 'react-quill/dist/quill.snow.css'
 
-import { ArticleInterface } from '@/apis/articles'
+import { ArticleInterface } from '@/apis/articles/route'
 
 interface Props {
   contentHtml: ArticleInterface['content']['html']
